# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/CompraBonos/GetCompra": {
      "get": {
        "tags": [
          "CompraBonos"
        ],
        "operationId": "GetCompra",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBonoDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBonoDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBonoDetailsDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/CompraBonos/CreateCompra": {
      "post": {
        "tags": [
          "CompraBonos"
        ],
        "operationId": "CreateCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraBonoForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraBonoForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompraBonoForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBonoDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBonoDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBonoDetailsDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompraBonoDetailsDTO": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "apellido1": {
            "type": "string",
            "nullable": true
          },
          "apellido2": {
            "type": "string",
            "nullable": true
          },
          "metodoPago": {
            "$ref": "#/components/schemas/MetodoPago"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraBonoItemDTO"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CompraBonoForCreateDTO": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "apellido1": {
            "type": "string",
            "nullable": true
          },
          "apellido2": {
            "type": "string",
            "nullable": true
          },
          "metodoPago": {
            "$ref": "#/components/schemas/MetodoPago"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraBonoItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraBonoItemDTO": {
        "required": [
          "cantidad"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "pvp": {
            "type": "number",
            "format": "float"
          },
          "nBocadillos": {
            "type": "integer",
            "format": "int32"
          },
          "nombreBono": {
            "type": "string",
            "nullable": true
          },
          "tipoBocadillo": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "MetodoPago": {
        "enum": [
          "Tarjeta",
          "Paypal",
          "Gpay"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}